requirements.txt
# Web Scraping
selenium==4.16.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Telegram
python-telegram-bot==20.7

# Environment & Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Utilities
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.12.1
flake8==6.1.0
mypy==1.7.1
ruff==0.1.8

# Type stubs
types-python-dateutil==2.8.19.14
